requirements.txt
streamlit
music21
langchain
python-dotenv
langchain_core
langchain_community
langchain_groq
scipy
synthesizer